changed properties to take float instead of int, created a new type of enemy, improved player movement

+++ src/main/java/main/BasicEnemy.java
@@ -6,7 +6,7 @@ public class BasicEnemy extends GameObject {
 
     private Handler handler;
 
-    public BasicEnemy(int x, int y, ID id, Handler handler) {
+    public BasicEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
         velX = 5;
         velY = 5;
@@ -32,11 +32,11 @@ public class BasicEnemy extends GameObject {
     @Override
     public void render(Graphics g) {
         g.setColor(Color.RED);
-        g.fillRect(x,y,16,16);
+        g.fillRect((int)x,(int)y,16,16);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16,16);
+        return new Rectangle((int)x, (int)y, 16,16);
     }
 }
