+++ src/main/java/main/GameObject.java
@@ -3,13 +3,13 @@ package main;
 import java.awt.*;
 
 public abstract class GameObject {
-    protected int x;
-    protected int y;
+    protected float x;
+    protected float y;
     protected ID id;
-    protected int velX;
-    protected int velY;
+    protected float velX;
+    protected float velY;
 
-    public GameObject(int x, int y, ID id) {
+    public GameObject(float x, float y, ID id) {
         this.x = x;
         this.y = y;
         this.id = id;
@@ -27,7 +27,7 @@ public abstract class GameObject {
         this.id = id;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
@@ -35,7 +35,7 @@ public abstract class GameObject {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
@@ -43,7 +43,7 @@ public abstract class GameObject {
         this.y = y;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
@@ -51,7 +51,7 @@ public abstract class GameObject {
         this.velX = velX;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
