+++ src/main/java/main/Trail.java
@@ -10,7 +10,7 @@ public class Trail extends GameObject {
     private int height;
     private float life;
 
-    public Trail(int x, int y, ID id, Color color, int width, int height, float life, Handler handler) {
+    public Trail(float x, float y, ID id, Color color, int width, int height, float life, Handler handler) {
         super(x, y, id);
         this.handler = handler;
         this.color = color;
@@ -35,7 +35,7 @@ public class Trail extends GameObject {
         g2d.setComposite(makeTransparent(alpha));
 
         g.setColor(color);
-        g.fillRect(x, y, width, height);
+        g.fillRect((int)x, (int)y, width, height);
 
         g2d.setComposite(makeTransparent(1));
     }
